Carve `TestSparseTensors` and `TestEinsum` out of the monolith (wala/ML#635).
The 19 sparse-tensor and 19 `tf.einsum` tests move verbatim into their
classes; the monolith keeps 752, preserving the 1,040 total.
